fix: 止损用HIGH/LOW检查intrabar穿越 + 跨方向冷却 + SPOT_SELL方向修正
P0: 硬止损/常规止损改用bar HIGH(空仓)/LOW(多仓)检查intrabar价格穿越,
    成交价限定在止损触发价+滑点, 防止实际亏损远超预设(-50% vs -28%)
P1: 强平检查同样使用HIGH/LOW, 防止错过intrabar强平事件
P1: 止损/硬止损/强平后设置跨方向冷却(cooldown*3), 防止同bar反向开仓
fix: SPOT_SELL direction从'long'改为'short', 修正前端方向显示

+++ strategy_futures.py
@@ -286,7 +286,7 @@ class FuturesEngine:
         self.spot_eth -= qty
         self.total_spot_fees += fee
         self.total_slippage_cost += slippage_cost
-        self._record_trade(dt, price, 'SPOT_SELL', 'long', qty, revenue, fee, 1, reason,
+        self._record_trade(dt, price, 'SPOT_SELL', 'short', qty, revenue, fee, 1, reason,
                            exec_price=actual_p, slippage_cost=slippage_cost,
                            pnl=real_pnl)  # 修正: 记录真实已实现PnL(成本法)
